examples/gemini/implementation/requirements.txt
ciaf-watermarks>=0.1.0
google-generativeai>=0.3.0
